#154, refactor domain class groups as Kotlin files.
